Wire up the observable into streaming requests to the Foyle service.
